encoder/requirements.txt
grpcio
grpcio-tools
sentence-transformers